Drag-to-dismiss FileViewController
I attached the gesture recognizer to the entire thumbnail view to be more intuitive and easier to grab for the user. Maybe we should remove the arrow button as well?

I had to do some yucky things to get the mini player and the header bar to play nice with the animation, but I think the end result is pretty good!