Only log JSON (and form JSON strings for logging) during debug mode, with a switch
Before this change, in release mode all 3 print statements did the work of forming String objects, even in release mode. The print in Lbry.swift also printed the JSON to stdout. It's a lot of work.

After this change, JSON responses are not stringified or printed unless debugging, and unless the specified line is uncommented.

Let me know if the default should be for JSON to be printed during debugging.