LbryUri: use NSString.substring(with:) for getting regex groups
NSRegularExpression uses (the equivalent of) unicodeScalars, but the
swift Range(_,in:) function operates on characters.

Use NSString.substring(with:) instead as it operates on unicodeScalars.

Fix crash when a zero-width joining character like ் is the only
character in a URI component.

Fix: #279
Fix: #429
Supersedes: #394 (Still keep this around as switching to Swift 5.7 Regex
simplifies a lot of code)